starter-kits/llamaindex/requirements.txt
llama-index-core>=0.10.0
llama-index-llms-openai>=0.1.0
streamlit>=1.35.0
python-dotenv>=1.0.0
